feat: update Order model and migrations to use bigInteger for price and transaction_amount, add casts for status and transaction_amount

+++ database/migrations/2025_11_27_031143_create_orders_table.php
@@ -19,7 +19,7 @@ return new class extends Migration {
             ); // Se apagar o produto, mantemos o pedido por histórico (sem cascade)
 
             // Dados Financeiros
-            $table->decimal('transaction_amount', 10, 2);
+            $table->bigInteger('transaction_amount');
             $table->string('payment_method')->nullable(); // pix, credit_card
 
             // Controle de Estado e Integração
